requirements.txt
numpy==1.19.5
plotly==4.14.3
fuzzywuzzy==0.18.0
streamlit==0.88.0
yfinance==0.1.63
pandas==1.2.1
requests==2.25.1
alpha_vantage==2.3.1
beautifulsoup4==4.10.0
Pillow==8.3.2
